Continuing to change up the backend.
Added a lot of unittests to aid in validating/debugging the changes. Still haven't updated print_all_plaintexts and the graph functions, but nearly back to the functionality from the original release. While it was painful to make such drastic changes, I'm optimistic they will set the stage to make this framework much more extensible.