Added strikes and sessions. Also working on parsing JtR Logs
A lot of WIP, but there is a lot of progress! I expect Strikes and Sessions to change around a lot as I use them more. The JtR log parsing is probably 90% complete from my tests, which means it's likely 5% complete from all the ways it could show up in the real world.... Without going into the JtR codebase though I think it's ok to add new parsing as I find additional logs that are interesting.