Lots of additions. Getting ready for CMIYC2024
Finally just cut out a lot of the hashes and outputs from the CMIYC2022 writeup to keep the framework size down. Longer term I need to figure something else out. Maybe include writeups in a submodule vs this git directory? The big change is also adding in support for PGP e-mails without having to install gnupgp. The team creation portion seems to be working but I have a lot of work to incorporate this into submitting cracked hashes.